Create a application registry
